Remove goroutines from postgres driver and migrate lib
There was an annoying error when calling migrate from code.
When a migration is stuck, migrate was stuck in a goroutine, resulting in the whole program to not
react to ctrl+c anymore.

I'm not quite sure why pipes and goroutines are even needed, as the migration needs to be run
sequential anyways.

This patch removes the goroutines and simplifies the code.